extractValue/totalWeight/requirements.txt
cachetools==5.3.0
charset-normalizer==3.0.1
docopt==0.6.2
gcsfs==2023.1.0
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-functions==1.11.0
google-cloud-pubsub==2.15.0
google-cloud-storage==2.7.0
google-cloud-tasks==2.12.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
mkl-fft==1.3.1
mkl-service==2.4.0
pandas==1.3.5
pipreqs==0.4.11
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
rsa==4.9
urllib3==1.26.14
weighted-levenshtein==0.2.2
yarg==0.1.9
fsspec==2023.1.0
